Item quantity adjustment to allow nulls.

+++ EFCommands/Items/EFAddItemCommand.cs
@@ -34,7 +34,7 @@ namespace EFCommands.Items
                 isForSale = request.isForSale,
                 ItemTypeId = request.ItemTypeId,
                 ItemQualityId = request.ItemQualityId,
-                Quantity = request.Quantity
+                Quantity = (int)request.Quantity
             });
 
             Context.SaveChanges();
